Suggestion to address issue #163
Ability to indicate a physical page number: Some of the use cases for the original page number requirement did not relate to page numbers. I have moved them since they are more related to bookmarking. I have also changed the language for this requirement and moved it to Random Access to Content since Synchronization did not seem to be an accurate reflection of what it was. Synchronization implies something is automatically happening whereas this page numbering requirement and use cases described a more manual cross referencing.

Ability to bookmark a precise position in the publication: This requirement already existed, but it has been rephrased to be more explicitly related to bookmarking. Two of the use cases from page numbering have been moved here. The Aika use case is less clear and I would like input on this one.

I have also added another requirement for preserving reading state, which is related to this topic.